+++ ontimize/components/form/o-form.component.ts
@@ -1005,7 +1005,6 @@ export class OFormComponent implements OnInit, OnDestroy, CanComponentDeactivate
   }
 
   protected postCorrectInsert(result: any) {
-    console.log('[OFormComponent.postCorrectInsert]', result);
     this.snackBarService.open('MESSAGES.INSERTED', { icon: 'check_circle' });
   }
 
@@ -1077,7 +1076,6 @@ export class OFormComponent implements OnInit, OnDestroy, CanComponentDeactivate
   }
 
   protected postCorrectUpdate(result: any) {
-    console.log('[OFormComponent.postCorrectUpdate]', result);
     this.snackBarService.open('MESSAGES.SAVED', { icon: 'check_circle' });
   }
 
@@ -1107,7 +1105,6 @@ export class OFormComponent implements OnInit, OnDestroy, CanComponentDeactivate
   }
 
   protected postCorrectDelete(result: any) {
-    console.log('[OFormComponent.postCorrectDelete]', result);
     this.snackBarService.open('MESSAGES.DELETED', { icon: 'check_circle' });
   }